chore: comment out & move before removing `testutils`
